functions/requirements.txt
firebase_functions~=0.1.0
Pillow~=10.1.0